requirements.txt
fastapi==0.115.8
python-dotenv==1.0.1
SQLAlchemy==2.0.38
uvicorn==0.34.0
asyncpg==0.30.0